move iterator def to utils
