+++ kernel/kernel-api/src/main/java/io/delta/kernel/internal/util/Utils.java
@@ -17,6 +17,8 @@
 package io.delta.kernel.internal.util;
 
 import io.delta.kernel.annotation.Evolving;
+import io.delta.kernel.data.FilteredColumnarBatch;
+import io.delta.kernel.data.Row;
 import io.delta.kernel.utils.CloseableIterator;
 import java.io.IOException;
 import java.util.Iterator;
@@ -123,5 +125,49 @@ public class Utils {
     } catch (Throwable throwable) {
       // ignore
     }
+  }
+
+  // Utility class to support `intoRows` below
+  private static class FilteredBatchToRowIter implements CloseableIterator<Row> {
+    private final CloseableIterator<FilteredColumnarBatch> sourceBatches;
+    private CloseableIterator<Row> current;
+    private boolean isClosed = false;
+
+    FilteredBatchToRowIter(CloseableIterator<FilteredColumnarBatch> sourceBatches) {
+      this.sourceBatches = sourceBatches;
+    }
+
+    @Override
+    public boolean hasNext() {
+      if (isClosed) {
+        return false;
+      }
+      while ((current == null || !current.hasNext()) && sourceBatches.hasNext()) {
+        closeCloseables(current);
+        FilteredColumnarBatch next = sourceBatches.next();
+        current = next.getRows();
+      }
+      return current != null && current.hasNext();
+    }
+
+    @Override
+    public Row next() {
+      if (!hasNext()) {
+        throw new java.util.NoSuchElementException("No more rows available");
+      }
+      return current.next();
+    }
+
+    @Override
+    public void close() throws IOException {
+      isClosed = true;
+      closeCloseables(current, sourceBatches);
+    }
+  }
+
+  /** Convert a ClosableIterator<FilteredColumnarBatch> into a CloseableIterator<Row> */
+  public static CloseableIterator<Row> intoRows(
+      CloseableIterator<FilteredColumnarBatch> sourceBatches) {
+    return new FilteredBatchToRowIter(sourceBatches);
   }
 }
